pyproject.toml
[project]
name = "RestauranteBD"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.0",
    "psycopg[binary,pool]>=3.3.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
    "yoyo-migrations>=9.0.0",
    "email-validator>=2.3.0",
    "validate-docbr>=1.11.1",
    "faker>=38.2.0",
    "plotly>=6.5.0",
    "common",
    "db",
]

[dependency-groups]
dev = [
    "colorlog>=6.10.1",
    "mypy>=1.19.0",
    "pandas-stubs>=2.3.3.251201",
    "python-json-logger>=4.0.0",
    "ruff>=0.14.8",
]

[tool.uv.workspace]
members = ["packages/db"]

[tool.uv.sources]
db = { workspace = true }

[tool.mypy]
strict = true
